bareos: copy updated scripts from production system.

+++ bareos/keys-backup.sh
@@ -12,13 +12,16 @@ BACKUP_CONFIG_FILE="/etc/bareos/keys-backup.cfg"
 DIR_CONFIG_NAME="bareos-dir"
 DIR_CONFIG_PATH="/etc/bareos/bareos-sd.d/director"
 PASSWORD_FILE="/etc/bareos/backup/.backup-password"
+PUBKEY_FILE="/etc/bareos/backup/recipient.pub"
 BACKUP_NAME="lto-hwe-keys.gpg"
 BACKUP_PATH="/etc/bareos/backup"
 BACKUP_SCRIPT="/etc/bareos/scripts/hwe-backup.sh"
 DECRYPT_IS_CRITICAL="true"
 ETCKEEPER_AUTOCOMMIT="false"
+GPG_ENCRYPT_MODE="password"
 
 MAIL_ENABLE="false"
+MAIL_ALWAYS="false"
 MAIL_HOST="localhost"
 MAIL_FROM="(Bareos) <bareos@localhost>"
 MAIL_TO="root"
@@ -46,6 +49,8 @@ fi
 TIME=$(date '+%H:%M:%S%z')
 DATE=$(date '+%Y.%m.%d')
 TIMESTAMP=$(date '+%Y.%m.%d_%H:%M:%S%z')
+BACKUP_CHANGED="false"
+SHOULD_SEND_MAIL="false"
 
 if [ -e "${BACKUP_CONFIG_FILE}" ]; then
   source "${BACKUP_CONFIG_FILE}"
@@ -57,9 +62,40 @@ if [ ! -e "${CONFIG_PATH}" ]; then
   exit 1
 fi
 
-# NOTE: This password is used to encrypt the backup.
-if [ ! -e "${PASSWORD_FILE}" ]; then
+if [ ! -e "${BACKUP_SCRIPT}" ]; then
+  echo >&2 "ERROR: The specified backup script '${BACKUP_SCRIPT}' does not exist."
+  exit 1
+fi
+
+encryptModes=([1]=password [2]=publicKey)
+if [ -z "${GPG_ENCRYPT_MODE}" ]; then
+  echo >&2 "ERROR: No GPG encryption mode specified. Backup aborted."
+  exit 1
+fi
+
+# Validate selected encryption mode
+re='^[0-9]+$'
+temp=$GPG_ENCRYPT_MODE
+if [[ $temp =~ $re ]] ; then
+  temp="${encryptModes[temp]}"
+  if [ -z "${temp}" ]; then
+    echo >&2 "ERROR: '$GPG_ENCRYPT_MODE' is not mapped to a known encryption mode."
+    exit 1
+  fi
+  GPG_ENCRYPT_MODE="$temp"
+else
+  if ! echo ${encryptModes[@]} | grep -q -w "$temp"; then
+    echo >&2 "ERROR: '$GPG_ENCRYPT_MODE' is not a valid GPG encryption mode."
+    exit 1
+  fi
+fi
+
+# Validate password encryption settings
+if [ "${GPG_ENCRYPT_MODE}" == "password" ] && [ ! -e "${PASSWORD_FILE}" ]; then
   echo >&2 "ERROR: Password file '${PASSWORD_FILE}' does not exist."
+  exit 1
+elif [ "${GPG_ENCRYPT_MODE}" == "publicKey" ] && [ ! -e "${PUBKEY_FILE}" ]; then
+  echo >&2 "ERROR: GnuPG public-key file '${PUBKEY_FILE}' does not exist."
   exit 1
 fi
 
@@ -75,7 +111,11 @@ function panic()
 
   # Remove temporary home and exit.
   echo >&2 "ERROR: ${msg}"
-  rm -rf $GNUPGHOME
+
+  if [ -e "${GNUPGHOME}" ]; then
+    rm -rf $GNUPGHOME
+  fi
+
   exit 1
 }
 
@@ -83,45 +123,93 @@ function panic()
 FILE1="${BACKUP_PATH}/${BACKUP_NAME}"
 CURRENT=""
 DECRYPT_FAILED="false"
-if [ -e "${FILE1}" ]; then
+if [ "${GPG_ENCRYPT_MODE}" == "password" ] && [ -e "${FILE1}" ]; then
   CURRENT=$(echo "${GPG_PASSWORD}" | gpg --batch --no-options --passphrase-file="${PASSWORD_FILE}" --homedir="${GNUPGHOME}" --armor --decrypt "${FILE1}" 2>/dev/null)
   if ! [ $? -eq 0 ]; then
-    if [ "${DECRYPT_IS_CRITICAL}" == "false" ]; then
+    if [ "${DECRYPT_IS_CRITICAL}" != "false" ]; then
       panic "Failed to decrypt previous backup."
     else
       DECRYPT_FAILED="true"
     fi
   fi
+else
+  DECRYPT_FAILED="true"
 fi
 
 # Generate a new backup
 NEW=$(${BACKUP_SCRIPT} ${DIR_CONFIG_NAME})
+if ! [ $? -eq 0 ]; then
+  panic "Script '${BACKUP_SCRIPT}' returned a non-zero exit code; backup aborted."
+fi
 
 # Check for differences between backups
 HEAD_LINES=3
 DIFF=$(diff <( echo "${NEW}" | tail -n +${HEAD_LINES}) <(echo "${CURRENT}" | tail -n +${HEAD_LINES}))
 if [ ! -z "${DIFF}" ] || [ "${DECRYPT_FAILED}" == "true" ]; then
   # Backup has changed; update file
+  if [ "${DECRYPT_FAILED}" == "true" ]; then
+    echo >&2 "Failed to decrypt previous encryption keys backup; re-writing backup..."
+  elif [ ! -z "${DIFF}" ]; then
+    echo "Hardware encryption keys backup has changed; writing new backup..."
+  fi
   if [ -e "${FILE1}" ]; then
     rm "${FILE1}"
   fi
-  echo "${NEW}" | gpg --batch \
-                      --no-options \
-                      --no-emit-version \
-                      --no-comments \
-                      --passphrase-file="${PASSWORD_FILE}" \
-                      --homedir="${GNUPGHOME}" \
-                      --symmetric \
-                      --force-mdc \
-                      --armor \
-                      --s2k-cipher-algo aes256 \
-                      --s2k-digest-algo sha512 \
-                      --s2k-mode 3 \
-                      --s2k-count 65000000 \
-                      --output "${FILE1}" > /dev/null 2>&1
 
-  if ! [ $? -eq 0 ]; then
-    panic "Backup encryption failed; update aborted."
+  if [ "${GPG_ENCRYPT_MODE}" == "password" ]; then
+    echo "${NEW}" | gpg --batch \
+                        --no-options \
+                        --no-emit-version \
+                        --no-comments \
+                        --passphrase-file="${PASSWORD_FILE}" \
+                        --homedir="${GNUPGHOME}" \
+                        --symmetric \
+                        --force-mdc \
+                        --armor \
+                        --s2k-cipher-algo aes256 \
+                        --s2k-digest-algo sha512 \
+                        --s2k-mode 3 \
+                        --s2k-count 65000000 \
+                        --output "${FILE1}" > /dev/null 2>&1
+
+    if ! [ $? -eq 0 ]; then
+      panic "Backup password encryption failed; update aborted."
+    fi
+
+    BACKUP_CHANGED="true"
+  elif [ "${GPG_ENCRYPT_MODE}" == "publicKey" ]; then
+    KEY_ID=$(gpg --homedir "${GNUPGHOME}" --keyid-format 0xlong --import "${PUBKEY_FILE}" 2>&1 | grep -Po '(?<=gpg\:\skey\s)[a-zA-Z0-9]+(?=\:\spublic)')
+    if ! [ $? -eq 0 ]; then
+      echo >&2 "ERROR: Failed to import public key."
+      exit 1
+    fi
+
+    if [ -z "${KEY_ID}" ]; then
+      echo >&2 "ERROR: Failed to determine GnuPG key identifier."
+    else
+      echo "Using GnuPG public-key with ID ${KEY_ID} for backup encryption."
+    fi
+
+    echo "${NEW}" | gpg --batch \
+                        --encrypt \
+                        --trust-model always \
+                        --no-options \
+                        --no-emit-version \
+                        --no-comments \
+                        --digest-algo sha512 \
+                        --cipher-algo aes256 \
+                        --homedir="${GNUPGHOME}" \
+                        --armor \
+                        --recipient ${KEY_ID} \
+                        --output "${FILE1}" > /dev/null 2>&1
+
+    if ! [ $? -eq 0 ]; then
+      panic "Backup public-key encryption failed; update aborted."
+    fi
+
+    BACKUP_CHANGED="true"
+  else
+    panic "Unsupported GnuPG encryption mode '${GPG_ENCRYPT_MODE}'; update aborted."
   fi
 
   # git handling for etckeeper (check if /etc/.git exists)
@@ -132,26 +220,42 @@ if [ ! -z "${DIFF}" ] || [ "${DECRYPT_FAILED}" == "true" ]; then
         git add "${FILE1}"
         git commit -m "bareos: auto-commit updated backup."
         popd > /dev/null 2>&1
+        echo "Committed updated backup file to local /etc Git repository."
       fi
     fi
   fi
+else
+  echo >&2 "Encryption keys have not changed; backup not updated."
+fi
 
+if [ "${MAIL_ALWAYS}" == "true" ] || [ "${BACKUP_CHANGED}" == "true" ]; then
+  SHOULD_SEND_MAIL="true"
+fi
+
+if [ "${MAIL_ALWAYS}" == "true" ] && [ "${MAIL_ENABLE}" != "true" ]; then
+  echo >&2 "WARNING: E-mail not enabled but option to always send e-mails is. Check your settings."
+fi
+
+if [ "${MAIL_ENABLE}" == "true" ] && [ "${SHOULD_SEND_MAIL}" == "true" ]; then
   # Send the updated backup file via bsmtp
-  if [ "${MAIL_ENABLE}" == "true" ]; then
-    if [ ! -z "${MAIL_HEADER}" ]; then
-      printf "${MAIL_HEADER}\n\n$(cat ${FILE1})\n\n${MAIL_FOOTER}\n" | bsmtp -h "${MAIL_HOST}" -f "${MAIL_FROM}" -s "${MAIL_SUBJECT}" "${MAIL_TO}"
-    elif [ ! -z "${MAIL_FOOTER}" ]; then
-      printf "%s\n\n" "$(cat ${FILE1})" "${MAIL_FOOTER}" | bsmtp -h "${MAIL_HOST}" -f "${MAIL_FROM}" -s "${MAIL_SUBJECT}" "${MAIL_TO}"
-    else
-      cat "${FILE1}" | bsmtp -h "${MAIL_HOST}" -f "${MAIL_FROM}" -s "${MAIL_SUBJECT}" "${MAIL_TO}"
-    fi
-    if ! [ $? -eq 0 ]; then
-      panic "Failed to e-mail backup file."
-    fi
+  if [ ! -z "${MAIL_HEADER}" ]; then
+    printf "${MAIL_HEADER}\n\n$(cat ${FILE1})\n\n${MAIL_FOOTER}\n" | bsmtp -h "${MAIL_HOST}" -f "${MAIL_FROM}" -s "${MAIL_SUBJECT}" "${MAIL_TO}"
+  elif [ ! -z "${MAIL_FOOTER}" ]; then
+    printf "%s\n\n" "$(cat ${FILE1})" "${MAIL_FOOTER}" | bsmtp -h "${MAIL_HOST}" -f "${MAIL_FROM}" -s "${MAIL_SUBJECT}" "${MAIL_TO}"
+  else
+    cat "${FILE1}" | bsmtp -h "${MAIL_HOST}" -f "${MAIL_FROM}" -s "${MAIL_SUBJECT}" "${MAIL_TO}"
   fi
+  if ! [ $? -eq 0 ]; then
+    panic "Failed to e-mail hardware encryption keys backup file to ${MAIL_TO}"
+  fi
+  echo "Sent hardware encryption keys backup via e-mail to ${MAIL_TO}"
+else
+  echo >&2 "Skipped sending hardware encryption keys backup to ${MAIL_TO}"
 fi
 
 # Remove temporary home
-rm -rf $GNUPGHOME
+if [ -e "${GNUPGHOME}" ]; then
+  rm -rf ${GNUPGHOME}
+fi
 
 exit 0
